Note: cannot access env from the same script
When a script is setting workflow env via "mykey=value" >> "$GITHUB_ENV"
You won't be able to read it using ${{ env.mykey }} within that script.

Makes perfect sense, how would the workflow be able to do that?

So always use the local value (and set it!) and use >> "$GITHUB_ENV" to
export it.