Change force email to override target_hour
To simulate the action of a cron, we can override TARGET_HOUR

This will mean that the rest of the action will not care about
the input setting and behave as if it were at "go time"

This means if the email was sent in the current day, it will not
send due to override.

If you need to do that, delete the emailSent event for the today,
in the database.